modules/connectiondata-api/requirements.txt
grpcio==1.38.0
protobuf3=0.2.1